+++ packages/terra-enzyme-intl/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+* Fixed
+  * Fixed jest test failure caused by react-intl update.
+
 ## Unreleased
 
 * Fixed