+++ packages/terra-functional-testing/CHANGELOG.md
@@ -7,6 +7,7 @@
   * Added `useRemoteReferenceScreenshots` cli option for downloading reference screenshots from a remote site for screenshot comparisons.
   * Delete the `reference` screenshot directory when `useRemoteReferenceScreenshots` is true.
   * Added a check of the `BUILD_BRANCH` environment variable to determine if tests should pass regardless of image mismatch.
+  * Added load configurations for downloading screenshots.
 
 * Changed
   * Changed how we consume `BUILD_BRANCH` environment variable to match what the expected values of it are.